+++ chapter3/em.py
@@ -6,7 +6,7 @@
 #       extension: .py
 #       format_name: light
 #       format_version: '1.5'
-#       jupytext_version: 1.14.7
+#       jupytext_version: 1.15.2
 #   kernelspec:
 #     display_name: Python 3 (ipykernel)
 #     language: python
@@ -32,12 +32,15 @@
 # $$
 # \nabla \cdot D = \rho,
 # $$
+#
 # $$
 # \nabla \cdot B = 0,
 # $$
+#
 # $$
 # \nabla \times E = -\frac{\partial B}{\partial t},
 # $$
+#
 # $$
 # \nabla \times H = \frac{\partial D}{\partial t}+ J.
 # $$
@@ -63,6 +66,7 @@
 # $$
 #     - \nabla \cdot (\mu^{-1} \nabla A_z) = J_z \qquad \text{in } \mathbb{R}^2,\\
 # $$
+#
 # $$
 # \lim_{\vert(x,y)\vert\to \infty}A_z = 0.
 # $$
@@ -79,18 +83,20 @@
 # $$
 # a(A_z, v)=\int_\Omega \mu^{-1}\nabla A_z \cdot \nabla v ~\mathrm{d}x,
 # $$
+#
 # $$
 # L(v)=\int_\Omega J_z v~\mathrm{d} x.
 # $$
+#
 
 # ## Meshing a complex structure with subdomains
 #
 # We create the domain visualized in the cross section figure above using gmsh. Note that we are using the `gmsh.model.occ.fragment` commands to ensure that the boundaries of the wires are resolved in the mesh.
+#
 
 # +
 from dolfinx import default_scalar_type
-from dolfinx.fem import (dirichletbc, Expression, Function, FunctionSpace,
-                         VectorFunctionSpace, locate_dofs_topological)
+from dolfinx.fem import (dirichletbc, Expression, Function, functionspace, locate_dofs_topological)
 from dolfinx.fem.petsc import LinearProblem
 from dolfinx.io import XDMFFile
 from dolfinx.io.gmshio import model_to_mesh
@@ -199,17 +205,20 @@ if mesh_comm.rank == model_rank:
 # -
 
 # As in [the Navier-Stokes tutorial](../chapter2/ns_code2) we load the mesh directly into DOLFINx, without writing it to file.
+#
 
 mesh, ct, _ = model_to_mesh(gmsh.model, mesh_comm, model_rank, gdim=2)
 gmsh.finalize()
 
 # To inspect the mesh, we use Paraview, and obtain the following mesh
+#
 
 with XDMFFile(MPI.COMM_WORLD, "mt.xdmf", "w") as xdmf:
     xdmf.write_mesh(mesh)
     xdmf.write_meshtags(ct, mesh.geometry)
 
 # We can also visualize the subdommains using pyvista
+#
 
 pyvista.start_xvfb()
 plotter = pyvista.Plotter()
@@ -226,10 +235,11 @@ else:
 
 
 # Next, we define the discontinous functions for the permability $\mu$ and current $J_z$ using the `MeshTags` as in [Defining material parameters through subdomains](./subdomains)
+#
 
 # +
 
-Q = FunctionSpace(mesh, ("DG", 0))
+Q = functionspace(mesh, ("DG", 0))
 material_tags = np.unique(ct.values)
 mu = Function(Q)
 J = Function(Q)
@@ -254,9 +264,10 @@ for tag in material_tags:
 # In the code above, we have used a somewhat less extreme value for the magnetic permability of iron. This is to make the solution a little more interesting. It would otherwise be completely dominated by the field in the iron cylinder.
 #
 # We can now define the weak problem
+#
 
 # +
-V = FunctionSpace(mesh, ("Lagrange", 1))
+V = functionspace(mesh, ("Lagrange", 1))
 tdim = mesh.topology.dim
 facets = locate_entities_boundary(mesh, tdim - 1, lambda x: np.full(x.shape[1], True))
 dofs = locate_dofs_topological(V, tdim - 1, facets)
@@ -269,14 +280,16 @@ L = J * v * dx
 # -
 
 # We are now ready to solve the linear problem
+#
 
 A_z = Function(V)
 problem = LinearProblem(a, L, u=A_z, bcs=[bc])
 problem.solve()
 
 # As we have computed the magnetic potential, we can now compute the magnetic field, by setting `B=curl(A_z)`. Note that as we have chosen a function space of first order piecewise linear function to describe our potential, the curl of a function in this space is a discontinous zeroth order function (a function of cell-wise constants). We use `dolfinx.fem.Expression` to interpolate the curl into `W`.
+#
 
-W = VectorFunctionSpace(mesh, ("DG", 0))
+W = functionspace(mesh, ("DG", 0, (mesh.geometry.dim, )))
 B = Function(W)
 B_expr = Expression(as_vector((A_z.dx(1), -A_z.dx(0))), W.element.interpolation_points())
 B.interpolate(B_expr)
@@ -284,6 +297,7 @@ B.interpolate(B_expr)
 # Note that we used `ufl.as_vector` to interpret the `Python`-tuple `(A_z.dx(1), -A_z.dx(0))` as a vector in the unified form language (UFL).
 #
 # We now plot the magnetic potential $A_z$ and the magnetic field $B$. We start by creating a new plotter
+#
 
 # +
 plotter = pyvista.Plotter()
@@ -300,10 +314,12 @@ else:
 # -
 
 # ## Visualizing the magnetic field
+#
 # As the magnetic field is a piecewise constant vector field, we need create a custom plotting function.
 # We start by computing the midpoints of each cell, which is where we would like to visualize the cell-wise constant vector.
-# Next, we take the data from the function `B`, and  shape it to become a 3D vector.
+# Next, we take the data from the function `B`, and shape it to become a 3D vector.
 # We connect the vector field with the midpoint by using `pyvista.PolyData`.
+#
 
 # +
 plotter = pyvista.Plotter()
@@ -312,7 +328,7 @@ plotter.set_position([0, 0, 5])
 # We include ghosts cells as we access all degrees of freedom (including ghosts) on each process
 top_imap = mesh.topology.index_map(mesh.topology.dim)
 num_cells = top_imap.size_local + top_imap.num_ghosts
-midpoints = compute_midpoints(mesh, mesh.topology.dim, range(num_cells))
+midpoints = compute_midpoints(mesh, mesh.topology.dim, np.arange(num_cells, dtype=np.int32))
 
 num_dofs = W.dofmap.index_map.size_local + W.dofmap.index_map.num_ghosts
 assert (num_cells == num_dofs)
@@ -328,3 +344,6 @@ if not pyvista.OFF_SCREEN:
     plotter.show()
 else:
     B_fig = plotter.screenshot("B.png")
+# -
+
+